requirements.txt
numpy==2.1.2
pillow==10.4.0
